requirements.txt
numpy==1.23.5
pandas==1.5.3
Flask==2.3.2
gunicorn==20.1.0
googletrans==4.0.0-rc1
python-dotenv==1.0.0
requests==2.31.0


